Fix: eleminating circular dep. (ESLint)
