+++ src/utils/index.ts
@@ -1,6 +1,4 @@
 // @ts-nocheck
-import { addGoal, createGoal } from "@src/api/GoalsAPI";
-
 export const formatDate = () => {
   const newDate = new Date();
   return newDate;
@@ -24,24 +22,4 @@ export const getDates = (startDate: Date, stopDate: Date) => {
     currentDate.setDate(currentDate.getDate() + 1);
   }
   return dateArray;
-};
-
-const random = (min: number, max: number) => Math.floor(Math.random() * (max - min)) + min;
-
-export const createDummyGoals = async () => {
-  console.log("wait");
-  const dummyNames: string[] = ["Walk", "Gym", "Study", "Shopping", "Nap", "Code Reviews", "Algo Practice"];
-  dummyNames.map(async (goalName: string) => {
-    const dummyGoal = createGoal(
-      goalName,
-      true,
-      random(1, 4),
-      null,
-      null,
-      0,
-      -1
-    );
-    const id = await addGoal(dummyGoal);
-    return id;
-  });
 };
